+++ crates/cpu/src/lib.rs
@@ -1,11 +1,10 @@
 pub mod bus;
 pub mod cpu;
 pub(crate) mod executor;
-pub mod futures;
 pub mod interface;
 mod opcodes;
 pub mod registers;
 
-pub use crate::bus::RegisterBus;
+pub use crate::bus::Bus;
 pub use crate::cpu::Cpu;
-pub use crate::interface::{New, Registers};
+pub use crate::interface::Registers;
