+++ crates/cpu/src/opcodes/arithmetic.rs
@@ -1,7 +1,7 @@
 use crate::registers::{Bits16, Bits8, Arithmetic as A};
-use crate::bus::RegisterBus;
+use crate::bus::Bus;
 use crate::cpu::Registers;
-use crate::futures::{GetAt, NextPc};
+use crate::registers::futures::{GetAt, NextPc};
 use memory::Memory;
 use num_enum::TryFromPrimitive;
 
@@ -318,7 +318,7 @@ impl Arithmetic {
 mod test_arithmetic {
     use super::Arithmetic;
     use crate::registers::{Bits16, Bits8, Flag};
-    use crate::{executor, RegisterBus, Registers};
+    use crate::{executor, Bus, Registers};
     use memory::Memory;
 
     #[test]
