+++ crates/cpu/src/opcodes/rotate.rs
@@ -1,5 +1,5 @@
 use super::consts::{BIT0, BIT7};
-use crate::bus::RegisterBus;
+use crate::bus::Bus;
 use crate::opcodes::Src;
 use crate::{
     registers::{Bits8, Flag},
@@ -182,7 +182,7 @@ impl Rotate {
 mod test_rotate {
     use super::Rotate;
     use crate::registers::{Bits16, Bits8, Flag};
-    use crate::{executor, RegisterBus, Registers};
+    use crate::{executor, Bus, Registers};
     use memory::Memory;
 
     #[test]
