+++ crates/cpu/src/registers/logical.rs
@@ -1,6 +1,6 @@
 use super::{Bits8, Flag, Arithmetic};
 use crate::registers::Registers;
-use crate::RegisterBus;
+use crate::Bus;
 
 pub trait Logical<T> {
     fn and(&mut self, area: T);
